set RUNTIME_VERSION in skse64_loader_common.vcxproj
This patch sets the runtime version and target name for the v140 and v142
builds.

 target                      toolset  Runtime Version      Skyrim Version
--------------------------------------------------------------------------
 Debug, Release              v140     0x01050610 (1.5.97)  Skyrim SE
 Debug_VC142, Release_VC142  v142     0x01061430 (2.1.3)   Skyrim AE

The target name used to be configured in 'Common.props' but this is
problematic since the VC142 build uses a different runtime and the DLL's
filename needs to change accordingly. Because of this the parameter is
now set in the project file.